feat: provide function to reset GPSL1 decoder
The new function reset_decoder_state allows to reset a GPSL1 decoder. This will reset the bit buffer and TOW, while maintaining the remaining parameters. It can be useful for a GNSSReceiver to keep the usually still valid ephemeris data etc. in the decoder for a reacquisition of a temporally lost satellite.